This feels a bit better
